[SYCL][DeviceSanitizer] Checking out-of-bounds error on sycl::local_accessor (#13503)
UR: https://github.com/oneapi-src/unified-runtime/pull/1532

To check sycl::local_accessor(aka, dynamic local memory), we need to
extend a new argument in spir kernel, this is because:
- ASan needs to know some size information of local buffer, like its
size and size with redzone, so that it can poison its shadow memory
- By using this new argument, we can also pass some per-launch
information (that is, it is different in each launch of kernel). One
obvious example is SanitizerReport, which saves the error message, so
that we can store and print multiple error reports for one kernel with
different arguments. Another example is the shadow memory of local
memory, this should be different per-launch as well, since one kernel
can be launched multiple times and executed in parallel.

I named this argument as "__asan_launch", which is a pointer pointed to
"LaunchInfo" structure and allocated it in shared USM. To make this
pointer can be used in spir_func w/o extending their argument, I created
a global external local memory (external, so that it can be shared with
other translation units, and its instance is defined in libdevice), and
save the "__asan_launch" into this local memory immediately at the entry
of kernel.

UR can't check the name of kernel arguments, so it can't know if the
kernel has "__asan_launch". So I assume the "__asan_launch" is always
there, and added a check to prevent DAE pass from removing it.

---------

Co-authored-by: Maosu Zhao <maosu.zhao@intel.com>
Co-authored-by: Kenneth Benzie (Benie) <k.benzie@codeplay.com>