[SYCL] Device code generation for "free functions", a new way to define kernels (#13207)
This PR is intended for review of the device-side code generation of
"free functions".

The free function spec is here:
https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/proposed/sycl_ext_oneapi_free_function_kernels.asciidoc

This PR does not implement the full specification. It does however allow
free function markup and execution using a manual method.

Later PRs will add the remaining parts of the spec. This is the current
status:
- Free functions are supported at file scope only.
- The SYCL_EXTERNAL markup is needed for free functions in addition to
the SYCL_EXT_ONEAPI_FUNCTION_PROPERTY property.
- The traits described in the specification section "New traits for
kernel functions" are not yet implemented.
- The functions described in the specification section "New kernel
bundle member functions" are not yet implemented. As a result, a
free-function kernel can only be launched by finding the kernel_id that
matches the function's name, and this requires knowing how the compiler
mangles the function's name when creating the kernel. This is a
temporary solution until the "New kernel bundle member functions" are
implemented.
- The compiler does not yet diagnose an error if the application
violates any of the restrictions listed in the specification under the
section "Defining a free function kernel".
- Device code generation is supported for scalars and USM pointers only.
It is not supported for complex kernel argument types requiring
decomposition like accessor, local_accessor, or stream.
- The implementation has not been tested to handle the case when a
kernel argument is optimized away. The switch
-fno-sycl-dead-args-optimization could be used to disable this
optimization, if needed
- The kernel information descriptor info::kernel::num_args cannot yet be
used to query the number of arguments in a free function kernel.