[CI] New workflow to ping issue assignee
There is a set of "backlog" issues we'd like to handle. The new process
is to confirm the issue and add the `confirmed` label, so the issue will
be automatically copied to our system and then assigned to the
appropriate team.

Sometimes the issue has an assignee but doesn't get any updates for a long
time. We decided that in this case we'd like to occasionally ping the
assignee to progress the work on this issue. There is also a big
bunch of old issues with an assignee and no updates.