Use fence rather than event for sync in L0 command-buffer update
The Level Zero adapter implementing
`urCommandBufferUpdateKernelLaunchExp` is doing a blocking host wait
with `zeEventHostSynchronize` on the executing of the command-buffer.

However, there is no guarantee that the command-buffer has been
enqueued before update has been called. Resulting in a hang if this
is not the case. Even if a command-buffer has been enqueued before
the event can't guarantee that the command-list has finished executing.

Instead, replace the blocking wait on a L0 event with a blocking
wait on a L0 fence from the last submission of the command-buffer. This is
because fences are the only safe way to track when a command-buffer
execution has completed.