[SYCL] Implementation of preprocessing output when option '-include-footer' is used. (#13607)
Current implementation to support our integration header and footer
files during the host specific compilation, is to create a temporary
file based on the source file with the integration footer appended to
it. This temporary file is then compiled as the main source file.
Use of `ccache` doesn't work well with this implementation, as the
actual compiled source file does not match the source file given on the
command line.
To solve this, a new option named `-include-footer` is introduced. With
it, the integration footer is appended to the source file during
preprocessing.

Syntax : `-include-footer filename`
`filename`: represents the integration footer file name generated during
the device compilation.

Use: `clang -cc1 -include int-header.h -include-footer int-footer.h
main.cpp`

This patch implements the changes produced in the output of the
preprocessing phase. It only uses this option. The implementation of the
option will be done in a different patch.
The preprocessed output should include the name of the footer and its
content at the end of the source code.