[SYCL] Fix exception thrown for composite_devices query (#13868)
The sycl_ext_oneapi_composite_device extension specifies that the
composite_devices query should throw if a device returns false for
device::has with ext_oneapi_is_composite. However, the current
implementation simply returns an empty device list if the backend of the
device is not L0. This commit fixes this discrepancy.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>