[SYCL][Joint Matrix] Tests query for default subgroup size (#13798)
Changes:
* Main folder now does not specify the subgroup (SG) size , it queries
it
* SG32 folder overrides the default SG size, through a `#define`
* XMX8 folder does not specify the SG size, it queries it
* minor cleanups
* `joint_matrix_opt_kernel_feature_impl.hpp` - issues a different error
code than before. I believe now that it is a named kernel it might be
taking a different code path through the runtime. Was
`errc::kernel_not_supported` now `errc::invalid`. I checked and if i
modify the tile size of `N` to be reasonable the test passes as
expected.