[CTS] Fix UR_CONFORMANCE_ENABLE_MATCH_FILES=OFF
When configured without match files, the CTS was trying to run a
command and its arguments as a single command, as it was represented by
a single string variable. It needs to be a single string in the default
case - when building with UR_CONFORMANCE_ENABLE_MATCH_FILES=ON - as that
command string is passed off to another command via a flag, so it must
be seen by the shell as one item.

While definitely not *nice*, this is the simplest fix I could think of
which preserves the behaviour of the (more important, tested) 'default'
case.