Don't assume ctx contains all devs in platform
A context might only contain a subset of the devices in a platform. Each
ur_mem_handle_t_ contains some vectors of length ctx.NumDevices.
Device->getIndex() will return the index of a device in a platform.
Therefore, when we are using the ur_mem_handle_t vectors we must make
sure the ctxIdx doesn't exceed ctx.NumDevices.