[Driver][SYCL][NewOffloadModel] Improve arch association for device
When using the new offloading model, the values being passed to
-fsycl-targets were not fully being realized when attempting to target
specific architectures.  Uses of -fsycl-targets=intel_gpu_* were not
properly setting the arch values that are used down the line (namely the
packaging step).

Improve this situation by populating the assocated architecture mappings
with what is seen when parsing the -fsycl-targets option.  This applies
to all intel_gpu, nvidia_gpu and amd_gpu targets.