Change to calculating device index per ctx
The modulo operator is not sufficient to calculate the device Idx in a
context. It is necessary instead to find the index of a device in the
context's Devices vector.

We could also use a map for this, but since we use this ctx local idx
quite a lot, I was hesitant to transform all vectors to std::maps
because of possible perf implications of using a different data
structure.