[L0] Fix local build without system install
Use the include path for `ze_api.h` which points to the Level Zero
loader in the build dependencies, which does not have reside in the
`level_zero` subdirectory, in Level Zero adapter test suite.

I believe this passed GHA because of the presence of a system install of
Level Zero, rather than relying on the header acquired via fetch
content.