[SYCL][Bindless] Replace 'image_channel_order' field in 'image_descriptor' with number of channels (#13745)
'image_channel_order' field in 'image_descriptor' is and can only be
used for the number of channels due to CUDA having no notion of image
channel order. Replaced with the number of channels instead.

Also deprecate 'get_channel_order' function from 'image_mem' as it is
redundant because images no longer have or need a notion of channel
order. Only channel size.