[SYCL] Detach composite/component devices from L0
The current implementation of component and composite devices makes
multiple checks for the backend in order to bind the behavior of the
extension to L0. However, this should instead be done by the backends,
having them report missing support when possible, and allowing backends
to potentially support the extension in the future without needing
to change the SYCL runtime.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>