Add basic CUDA impl using ScopedStream
Use ScopedStream to return the same stream during the lifetime of the
RAII object. This allows us to create events outside a user submitted
func, and submit work within the user submitted func, since the stream
given to the user from urQueueGetNativeHandle is guaranteed to be the
same stream that we record events on.