[SYCL][XPTI] Fix off-by-one error in USMAnalyzer
When looking for the correct allocation, the upper bound check was
inclusive (Ptr <= Alloc.first + Alloc.second.Length). If we have two
allocations back-to-back, the pointer to the beginning of the second
allocation would incorrectly be determined as belonging to the first
allocation. This caused false-positives errors about out-of-bounds
memory operations.