Make unsafe atomic fadd opt in
Atomic fadd can fail for malloc_shared allocations if the PCIe bus
doesn't support PCIe atomics. This makes the default CAS atomics, so
that atomics are guaranteed to work. Fast, unsafe atomics can be used if
the flag -mllvm --amdgpu-oclc-unsafe-fp-atomics=true is used.

Testing is updated to test this new option.