Disable SYCL Device RUNs till we determine required behavior (#13907)
Using -fsycl-is-device does not make sense in these RUN lines since none
of the functions are kernels or device functions. I tried removing the
flag to verify if required IR is emitted but there is still address
space mismatch between what community clang emits and what we emit in
syclos. We have customizations in syclos for address spaces. It will
take some time to analyze and determine right steps forward. In the
meanwhile - disable these tests locally