[HIP] Various fixes for urEnqueueMemImage(Read|Write|Copy)
It appears that several width/height/depth parameters were mixed up.
This brings the HIP adapter in line with the CUDA adapter, and fixes
several UR CTS tests.

It also adds missing support for 1D image operations. It doesn't
optimize them in the same way that the CUDA adapater appears to (using
memcpy operations) but that can be left for future work.