[SYCL][E2E] Fix buffer_create by using L0 interop
The buffer_create test expects to know whether a device is integrated
based on a deprecated device info query. This was changed in good faith
to a seemingly equivalent aspect, but that aspect is not an exact
equivalent when it comes to checking for integrated devices.
This commit changes the test to instead use L0 interop to get
information about the device type directly from the L0 interfaces.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>