[SYCL] Force-emit more member functions into device code
We only emit unused member functions if they have been explicitly
annotated with `sycl_device` attribute (through `SYCL_EXTERNAL` macro).

SYCL extension for virtual functions introduces an alternative
markup for specifying which function and that markup is SYCL
compile-time properties that we turn into attributes implicitly under
the hood.

Essentially, we now have a situation where an implicit `sycl_device`
attribute on a member function should be treated as an explicit one,
because it could be a result of SYCL compile-time property being applied
to that method.