Set minimum L0 device support to support UR kernel update
In order to support all the ways a kernel command in a command-buffer
can be updated in Unified Runtime, the following capabilities need
to be available from the device:
* `ZE_MUTABLE_COMMAND_EXP_FLAG_KERNEL_ARGUMENTS`
* `ZE_MUTABLE_COMMAND_EXP_FLAG_GROUP_COUNT`
* `ZE_MUTABLE_COMMAND_EXP_FLAG_GROUP_SIZE`
* `ZE_MUTABLE_COMMAND_EXP_FLAG_GLOBAL_OFFSET`

Only supporting kernel argument update is not sufficient, as is
currently the case. There is also no way in the UR API currently
to take advantage of the `ZE_MUTABLE_COMMAND_EXP_FLAG_SIGNAL_EVENT` and
`ZE_MUTABLE_COMMAND_EXP_FLAG_WAIT_EVENTS` capabilities.