Remove L0 limitation from command-buffer update CTS tests
The command-buffer kernel update CTS tests currently assume that
L0 doesn't support updating the ND-Range configuration of kernels
in a command-buffer.

This is no longer the case, and the skips in these situations can
be removed. This patch also contains tweaks to the tests to address
issues hidden while the tests were skipped on Level Zero.