[SYCL][Graph] Fix PI Kernel leak in graph update
Running the `Graph/Update` E2E tests on Level Zero with
`UR_L0_LEAKS_DEBUG=1` shows that we are leaking a PI kernel
and module.

On investigation this was because we are retaining these objects
in `getOrCreateKernel()` but not releasing them. Added
release calls similar to how it is done in
[enqueueImpCommandBufferKernel](https://github.com/intel/llvm/blob/b49303c7e13ca0a69454eaaaeb8c3d094916218d/sycl/source/detail/scheduler/commands.cpp#L2550) by the scheduler