[HIP] Remove overly restrictive urEnqueueMemBufferFill validation
As noted in CUDA adapter commit 00761dc4, the validation for this entry
point was changed in commit 2c533e6a and hoisted into the validation
layer. The restriction on the pattern size being at most 128 was removed
in 00761dc4, and I can't see a reason for it being in the HIP adapter
save for the fact that it was copied from the CUDA adapter.