[SYCL] Annotate virtual calls in device IR
This commit is a part of virtual functions extension implementation that
is proposed in intel/llvm#10540.

As part of the feature, we would like to achieve both:
- allow virtual function calls from SYCL kernels that are submitted with
  the right properties as described in the language extension
  specification
- disallow virtual function calls from SYCL kernels that do not use any
  (or the right) properties to conform with the core SYCL 2020
  specification

Implementing such diagnostics will require call graph traversal to
understand the origin of the call and that is not what's easy to do in
FE. Therefore, the implementation design for virtual functions proposes
that such diagnostic is offloaded to a pass.

This commit is a preparations for it: it introduces an attribute to all
virtual function calls in device code so they can be easily detected and
not confused with plain function pointer calls (for which we don't have
a language specification and therefore well-defined behavior yet).