[Driver][SYCL][NewOffload] Update option passing for packager and AOT
When using the new offloading model, we need to be able to differentiate
the options that are being passed along to the packager as well as the
clang-linker-wrapper.

Update the behavior to pass any of the implied device compilation
options as well as the compile and link options into the packager so
they can be later processed by the clang-linker-wrapper.  The options
for the clang-linker-wrapper are also updated so users and update any
potential options that need to be used for the device compilation.