[SYCL][Joint Matrix] Test combinations are queried Part 4 (#14019)
Supported matrix dimensions are queried from the device, and inform the
tests which tile sizes one can use.

This is a subset of all tests that are planned to be modified.

Test manually tested on PVC and SPR - no new regresssions

The following tests have been marked as XFAIL on all platforms. I
removed them from XMX8 folder. Once they are passing then they can be
modified to query the supported matrix dimensions form the device.
* joint_matrix_colA_rowB_colC.cpp
* joint_matrix_out_bounds.cpp
* joint_matrix_unaligned_k.cpp