[SYCL] Re-implement diagnostics about virtual calls
With `sycl_ext_oneapi_virtual_functions` extensions we would like to
allow certain kernels to perform virtual function calls. That is if they
were submitted with the right properties.

That means that instead of simply checking for presence of virtual
function calls in device code, we need to analyze call chain to see how
exactly a kernel performing such call is defined.

This is not a task for the front-end and therefore the diagnostics
mechanism is moved to a pass, as suggested by the implementation design
proposed in intel/llvm#10540