Fix lifetime of pointer used in update
When we initialize our `ze_mutable_group_count_exp_desc_t` struct that's
used in a pointer chain passed to
`zexCommandListUpdateMutableCommandsExp` we set the
`pGroupCount` member to `&ZeThreadGroupDimensions`.

However, `ZeThreadGroupDimensions` is declared on the stack in the
scope of the if statement. Resulting in this pointer becoming invalid
after the scope of the if statement, when the L0 update API is called.

Fixed by moving the `ZeThreadGroupDimensions` variable outside the
scope of the if statement, into the scope of the function, so that
the address is valid when `zexCommandListUpdateMutableCommandsExp`
is called.