[GHA] Temporarily disable testing on windows-2022
As noted in #1715 testing on Windows Server 2022 in Release mode has
become flaky in the last couple of days. The root cause of this has so
far been elusive and the issue is not reproducible on a local Windows 11
machine. This patch temporarily disables this axis from the job matrix
while investigation continues.