[SYCL][Graph] Missing test dependency after data init
In the `Graph/Update/whole_update_dynamic_param.cpp` test there
is no dependency defined on the out-of-order queue between initalizing
the data and running the graph which uses it.

On immediate command-list L0 backends in particular this manifests
as a fail.

Fixed by adding a wait between data initialization and use.