[SYCL] Use kernel bundles to query reduction work-group sizes
Work-group sizes currently rely on device maximum rather than the max from a kernel query.
This changes aims to allow querying the kernels by using kernel bundles in the reduction implementations.
In order to do that, all kernel functions have to be defined as function objects and given unique names
(for the `kernel_id`s), so they can be identified for a kernel_bundle to be obtained.

Additionally, the kernel_bundle can be used with the command group execution to ensure we exeucte the
same kernel we queried. This has been added as an optional and can be user-configured via
`SYCL_REDUCTION_ENABLE_USE_KERNEL_BUNDLES=1|0`.
It is set to 0 by default, which means `use_kernel_bundle` is not called.

The goal of this is to be do `kernel.get_info<kernel_device_specific>` (i.e. `work_group_size`) queries,
primarily to make sure that safe work-group sizes are chosen, but also for other capability queries which
can be use for the implementations of the different kernel startegies.