[NFCI][SYCL] Move SYCL Module Splitting to library. Part 2 (#13282)
Added SYCL Module Splitting as a library. ESIMD splitting is not present
in this patch and will be added in an upcoming patch.
Added particular testing tool sycl-module-split that invokes added
functionality.

Not all `device-code-split` tests were updated in this patch because the rest
of them (mostly) don't test module splitting itself. They will be migrated in an
upcoming patch.