[SYCL] Disable in-order queue barrier optimization while profiling
Current implementation of profiling info for NOP barriers is inconsistent
with other events from the same queue. To make them consistent while keeping the optimization, we would
need to duplicate the event on our side and make the duplicate also check
and potentially use profiling information of its previous event.

Instead, as the first step, disable the NOP optimization during profiling
since profiling is known to incur a performance hit anyway. The proper
duplicate event approach can be implemented as a follow up if this causes issues for
users.