[SYCL][COMPAT] Add bfe_safe and bfi_safe APIs (#14006)
This PR adds bit-field extract (`bfe_safe`) and bit-field insert
(`bfi_safe`) to the `math.hpp` header.

These are 'bounds checked' variants of the `detail::bfe` and
`detail::bfi` APIs respectively, though in addition to bounds checking
the `_safe` variants also provide:
 - asm for NVPTX
 - Proper treatment of signed types (`bfe_safe`)

As such, it's not clear whether the 'unsafe' variants ought to be
exposed at all and so I've put them in `detail::` for now. What are the
expected semantics in relation to the `_safe` variants? They would
likely need separate tests, and it's not clear that DPCT use these.

---------

Signed-off-by: Joe Todd <joe.todd@codeplay.com>
Co-authored-by: Alberto Cabrera Pérez <alberto.cabrera@intel.com>