[Driver][SYCL][NewOffload] Fix duplication of device targets
When passing along multiple targets in the form of
-fsycl-targets=intel_gpu_dg1,intel_gpu_pvc, the number of the device
compilations was n*n as opposed to just n.  Due to how we were handling
duplicate entries for toolchain generation, the different names used
even though they had the same target triple (spir64_gen) we being
considered as unique, causing the multiple entries.

This is the second attempt to push this one in, updated the
sycl-offload-new-driver.c test to reflect ordering issues encountered.