[New offload driver][Device lib] Add SYCL device library files for all targets (#14102)
clang-linker-wrapper is not target-specific. i.e. it is not called for a
single target device. It is called only once.
Currently, clang-linker-wrapper is called only with device images with
spir64 targets. So, the existing approach to capture the first target
triple in the list of triples and use it for gathering
sycl-device-library files is valid. As we plan to add support for more
targets (AOT), we need to gather sycl-device-libraries for all targets.
This PR addresses this change.

Also, the triple should not be passed to the linker wrapper. The linker
wrapper should get the triples from device images.

Thanks

---------

Signed-off-by: Arvind Sudarsanam <arvind.sudarsanam@intel.com>