[SYCL][Graph] Update L0 aspect test (#14093)
The `Graph/UnsupportedDevice/device_query.cpp` test asserts that L0
devices will never have full graph support. This is not the case,
depending on the L0 device and driver version full graphs support is
possible.

Update the test to remove asserting on this, as diving into these
details is out of the scope of the test. This was previously decided
when discussion how to check the OpenCL backend for similar possible
variances in aspect support.