[CTS] Fix queue-releated nullptr issues in tests
These queue handles (i.e., pointers) weren't being initialized to
nullptr in the constructor and were in an indeterminate state. If they
weren't initialized as intended in the tests' SetUp routine (e.g.,
something else aborting the test beforehand), the queues were still in
an indeterminate state and possibly non-null, in which case the test
would try to free the garbage queues in TearDown.

This was responsible for one segfault in the Native CPU tests (which is
why we need to add a lot of new {{OPT}} failures, now that it may run to
completion) and for some HIP tests in an unrelated PR.