[SYCL] fix for tensorflow, linux only (#14153)
On Linux, the TensorFlow library (which uses SYCL .so) is having a
problem wherein its own releasing of resource is running afoul ours. The
solution is to stop deferring any new memory releases once we know we
are shutting down. To that end I've broken the shutdown into two parts
(early and late) and added a flag for whether its OK to defer memory or
not. This has been confirmed to fix the problem.
It is not entirely clear to me what, exactly, the underlying cause of
the problem is on the TensorFlow side. But I realized that while SYCL is
used by several intermediate libs we have no testing of this scenario,
so I've begun to add some.
Lastly, on Windows, I've got encouraging results that this approach may
help us unify shutdown so it's the same on both Linux and Win, but that
is a larger work and should be in its own PR.