[Driver][SYCL] Force an extension for AOT tool output files
When performing AOT and using -save-temps, make sure that the output
files are created with an extension.  This is done for 'ocloc' behaviors
as if an -output value does not have an extension, 'ocloc' will create a
file with a default '.bin' extension causing issues with file
consumption down the line.