[UR] Add interop testing for CUDA (#13988)
Add a basic interop test for CUDA, as well as an interop test that
ensures that memory migration will occur between devices in the same
context, when using host tasks. This is currently broken for CUDA but a
fix
is TODO. This fix is nontrivial so will take some time.