[SYCL][Doc] Update work-group-specific extension (#14271)
Several changes to the specification for this proposed extension:

* Change the name to `work_group_static`. This name is more consistent
with our other extension named `work_group_memory`.

* Change the phrase "work-group-specific memory" to "device local
memory". This is the same phrase the core SYCL spec uses to describe
this memory, and the word "specific" no longer makes sense now that we
changed the name of the extension.

* Change the constraint for the `T` template parameter. We cannot simply
require `T` to be trivially constructible / destructible because this is
not true for an unbounded array type.

* The restrictions about the legal scopes for a `work_group_static`
variable seem confusing. I'm not sure what we intended, so I just added
an issue for now.