[SYCL][Graph] Refine barrier semantics
Changes the semantics of
[sycl_ext_oneapi_enqueue_barrier](https://github.com/intel/llvm/blob/sycl/sycl/doc/extensions/supported/sycl_ext_oneapi_enqueue_barrier.asciidoc) commands recorded from a queue into a
SYCL-Graph.

The current semantics are that a barrier without a wait-list will
depend on all leaf nodes in the graph, and be a dependency of any
subsequent node added to the graph.

After discussion with users, this change updates the behavior of
barriers to only depend on leaf nodes from the same recording queue,
and only commands recorded from the same queue will have a dependency
on the barrier.