[SYCL][E2E] Reenable in_order_profiling_queue for L0
Due to some confusion about the output from the in_order_profiling_queue
test on L0, the test was disabled. However, the test can be safely
reenabled for that target, while keeping it disabled for FPGA.

Additionally, the failure in profiling_queue is believed to be due to
the same issue, so the JIRA has been added to it and the note in
in_order_profiling_queue has been updated to reflect the known
information about the failure.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>