[SYCL] Simplify storePlainArg to avoid alias violations (#14344)
The helper function storePlainArg in handler and jit_compiler
reinterpret-casts pointers to memory in vectors of char in order to
store arguments in them. However, this violates strict aliasing and is
unnecessary as the resulting pointers are immediately converted to void*
after all calls to the function. As such, this patch simplfies these
implementations to always return void* and use memcpy to avoid the alias
violation.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>