[SYCL] Fix GroupNonUniform builtin definitions
The implementation currently defines the GroupNonUniform SPIR-V builtins
with the same extension suffixes as their Group counterparts. However,
some of the Group counterparts are KHR extension functions, while the
GroupNonUniform operations are core SPIR-V operations. As such, the
SYCL-CTS tests fail to compile the non-uniform group extension tests
in full-conformance mode. This commit fixes this issue.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>