Get rid of post-commit testing
This is an experiment to  try and find a good balance of reliable test
infrastructure that can't be ignored during merge, tests enough, and
runs quickly.

It's based on the following principle:
Pull request testing should ask and answer the question

> If merged, would the proposed change represent a regression to the
project?

It should answer it to a high confidence, in as short a time as possible
It should not:
- test things that are unrelated to the user's code
- rely on dynamic data or unstable data beyond the author's control
  (e.g. last night's build results)
- be obviously incomplete

With the above principles in mind the jobs here are optimized to
maximise diversity coverage on available runners, while also minimizing
the turnaround time to know whether a merge request is not a regression.
There is a conflict here: high coverage relies on lots of hardware, and
takes lots of time, but since a long turnaround time discourages
contribution we have to make a very fine judgment of what is reasonable.
Here we do this by building *once* per native platform toolchain (with
potentially differenct configurations to broaden *build* coverage, and
afterwards running subsets of tests on each host's available adaptors

Mac OS = Apple Clang + libc++; DSO-linkage release build + assertions
Win32 = MSVC; release build - no assertions
Linux = gcc + libstdc++ release build + assertions

This allows us to avoid building with both clang and gcc on a Linux
platform - where the coverage gained is marginal (clang has slightly
different diagnostics but otherwise doesn't really exercise the input
code in a different manner to gcc) - but building with clang on Mac and
gcc on Linux does carry a more significant chance of identifying a build
regression. Since by default we enable `-Werror` in configuration, this
allows us also to minimize the number of warnings merged, but still
avoid doing multiple builds per platform.

In summary we're not really interested in testing the system compiler,
but instead ensuring that a reasonable matrix of configurations (static
vs dynamic vs noassert vs assert) do not regress.

We hopefully then answer the original question *reasonable confidence in
correctness* without the overhead of a brute-force combinations matrix.

We make the following assumptions:
- the system compiler and standard library are sane. It's tempting to
  run the end to end tests with multiple host compilers, but divergence
  in results here most likely indicates a host compiler bug, or bad
  behaviour in the runtime, which would show up regardless of host
  codegen. So when the end to end tests fail, they fail due to the
  user's code, rather than a test of the system
compiler.

Dogfooding is a noble goal, but it's not suitable for merge request
testing, since it introduces an instability for which the author of the
PR is *not responsible*. Thus: we hoist dogfooding into the nightly
testing which doesn't affect a user's chance of reasonably getting a
patch merged before the heat death of the universe.