[SYCL] interim kernel compiler with SYCL support (#14172)
SYCL language support on the part of the kernel_compiler is specified
here: https://github.com/intel/llvm/pull/11985

However, that specification is not presently realizable. We need more
support from the FE and post link tool to get the demangled names
propagated through. But it is usable before that, with constraints about
using extern "C" or knowing the mangled kernel name. We have folk that
want to test in the interim. I've refrained from updating the spec, and
instead this interim support in our experimental extension will be snuck
in until it can be completed in full - then we'll update the spec to
release it properly.